+++ api/service/post/post_service.go
@@ -1,15 +1,13 @@
 package service
 
 import (
-	"errors"
-	"fmt"
 	"log"
 	"math/rand"
 	"strconv"
 	"sync"
 
 	"prototype2/domain"
-	"prototype2/responses"
+	"prototype2/errors"
 )
 
 var once sync.Once
@@ -36,15 +34,15 @@ func NewPostService(r domain.PostRepository) domain.PostService {
 func (*postService) Validate(post *domain.Post) error {
 	log.Print("[PostService]...Validate")
 	if post == nil {
-		err := errors.New("The post is empty")
+		err := errors.BadRequest.New("The post is empty")
 		return err
 	}
 	if post.Title == "" {
-		err := errors.New("The post title is empty")
+		err := errors.BadRequest.New("The post title is empty")
 		return err
 	}
 	if post.Text == "" {
-		err := errors.New("The post text is empty")
+		err := errors.BadRequest.New("The post text is empty")
 		return err
 	}
 	return nil
@@ -66,8 +64,8 @@ func (p *postService) GetByID(idString string) (*domain.Post, error) {
 
 	id, err := strconv.ParseInt(idString, 10, 64)
 	if err != nil {
-		    err = responses.BadRequest.Wrapf(err, "interactor converting id to int")
-		    err = responses.AddErrorContext(err, "id", "wrong id format")
+		err = errors.BadRequest.Wrapf(err, "interactor converting id to int")
+		err = errors.AddErrorContext(err, "id", "wrong id format")
 
 		return nil, err
 	}
@@ -78,12 +76,15 @@ func (p *postService) GetByID(idString string) (*domain.Post, error) {
 func (p *postService) Delete(idString string) error {
 	log.Print("[PostService]...Delete")
 
-	n, err := strconv.ParseInt(idString, 10, 64)
-	if err == nil {
-		fmt.Printf("%d of type %T", n, n)
+	id, err := strconv.ParseInt(idString, 10, 64)
+	if err != nil {
+		err = errors.BadRequest.Wrapf(err, "interactor converting id to int")
+		err = errors.AddErrorContext(err, "id", "wrong id format")
+
+		return err
 	}
 
-	post, err := p.repo.FindByID(n)
+	post, err := p.repo.FindByID(id)
 	if err != nil {
 		return err
 	}
