+++ log/access.log
@@ -229,3 +229,35 @@
 2020/07/16 23:52:21 [PostController]...GetPosts
 2020/07/16 23:52:21 [PostService]...FindAll
 2020/07/16 23:52:21 [PostRepository]...FindAll
+2020/07/17 11:19:10 [PostRepository]...Migrate
+2020/07/17 11:19:10 [UserRepository]...Migrate
+2020/07/20 10:20:13 [PostRepository]...Migrate
+2020/07/20 10:20:13 [UserRepository]...Migrate
+2020/07/20 10:21:51 [PostController]...GetPost
+2020/07/20 10:21:51 [PostService]...GetByID
+2020/07/20 10:22:10 [PostController]...GetPost
+2020/07/20 10:22:10 [PostService]...GetByID
+2020/07/20 10:22:10 [PostRepository]...FindAll
+2020/07/20 10:27:05 [PostRepository]...Migrate
+2020/07/20 10:27:05 [UserRepository]...Migrate
+2020/07/20 10:27:09 [PostController]...GetPost
+2020/07/20 10:27:09 [PostService]...GetByID
+2020/07/20 10:27:09 [PostRepository]...FindAll
+2020/07/20 10:28:28 [PostRepository]...Migrate
+2020/07/20 10:28:28 [UserRepository]...Migrate
+2020/07/20 10:49:15 [PostRepository]...Migrate
+2020/07/20 10:49:15 [UserRepository]...Migrate
+2020/07/20 10:49:25 [PostController]...DeletePost
+2020/07/20 10:49:25 [PostService]...Delete
+2020/07/20 10:49:25 [PostRepository]...FindAll
+2020/07/20 10:52:40 [PostRepository]...Migrate
+2020/07/20 10:52:40 [UserRepository]...Migrate
+2020/07/20 10:53:58 [PostRepository]...Migrate
+2020/07/20 10:53:58 [UserRepository]...Migrate
+2020/07/20 11:22:48 [PostRepository]...Migrate
+2020/07/20 11:22:48 [UserRepository]...Migrate
+2020/07/20 11:42:34 [PostController]...DeletePost
+2020/07/20 11:42:34 [PostService]...Delete
+2020/07/20 11:42:34 [PostRepository]...FindAll
+2020/07/20 11:44:47 [PostRepository]...Migrate
+2020/07/20 11:44:47 [UserRepository]...Migrate
